\name{engelberg}
\docType{data}
\alias{engelberg}
\title{Monthly temperature and precipitation values at Engelberg station, Switzerland}
\description{
  A dataset containing monthly homogenized temperature and precipitation values recorded at the Engelberg meteorological station in Switzerland (Altitude: 1036 m; Coordinates: 46.8175N / 8.4017E).

Reference date of homogenization: Temperature: 01.12.2009; Precipitation: 01.12.2007

The variables are as follows:

\itemize{
 \item Year. Year of the record (1864--2015)
 \item Month. Month of the record (1--12)
 \item Temperature. Temperature in degrees C
 \item Precipitation. Precipitation in mm
}
}

\usage{data(engelberg)}
\format{A data frame with 1819 rows and 4 variables.}
\source{Federal Office of Meteorology and Climatology MeteoSwiss}
\keyword{datasets}
